# wazevo(frontend): faster non-imported global access