--- a/internal/engine/wazevo/frontend/frontend_test.go
+++ b/internal/engine/wazevo/frontend/frontend_test.go
@@ -1181,15 +1181,11 @@ blk0: (exec_ctx:i64, module_ctx:i64, v2:i32)
 			m:    testcases.GlobalsGet.Module,
 			exp: `
 blk0: (exec_ctx:i64, module_ctx:i64)
-	v2:i64 = Load module_ctx, 0x8
-	v3:i32 = Load v2, 0x8
-	v4:i64 = Load module_ctx, 0x10
-	v5:i64 = Load v4, 0x8
-	v6:i64 = Load module_ctx, 0x18
-	v7:f32 = Load v6, 0x8
-	v8:i64 = Load module_ctx, 0x20
-	v9:f64 = Load v8, 0x8
-	Jump blk_ret, v3, v5, v7, v9
+	v2:i32 = Load module_ctx, 0x8
+	v3:i64 = Load module_ctx, 0x18
+	v4:f32 = Load module_ctx, 0x28
+	v5:f64 = Load module_ctx, 0x38
+	Jump blk_ret, v2, v3, v4, v5
 `,
 		},
 		{
@@ -1198,18 +1194,14 @@ blk0: (exec_ctx:i64, module_ctx:i64)
 			exp: `
 blk0: (exec_ctx:i64, module_ctx:i64)
 	v2:i32 = Iconst_32 0x1
-	v3:i64 = Load module_ctx, 0x8
-	Store v2, v3, 0x8
-	v4:i64 = Iconst_64 0x2
-	v5:i64 = Load module_ctx, 0x10
-	Store v4, v5, 0x8
-	v6:f32 = F32const 3.000000
-	v7:i64 = Load module_ctx, 0x18
-	Store v6, v7, 0x8
-	v8:f64 = F64const 4.000000
-	v9:i64 = Load module_ctx, 0x20
-	Store v8, v9, 0x8
-	Jump blk_ret, v2, v4, v6, v8
+	Store v2, module_ctx, 0x8
+	v3:i64 = Iconst_64 0x2
+	Store v3, module_ctx, 0x18
+	v4:f32 = F32const 3.000000
+	Store v4, module_ctx, 0x28
+	v5:f64 = F64const 4.000000
+	Store v5, module_ctx, 0x38
+	Jump blk_ret, v2, v3, v4, v5
 `,
 		},
 		{
@@ -1220,50 +1212,34 @@ signatures:
 	sig1: i64i64_v
 
 blk0: (exec_ctx:i64, module_ctx:i64)
-	v2:i64 = Load module_ctx, 0x8
-	v3:i32 = Load v2, 0x8
-	v4:i64 = Load module_ctx, 0x10
-	v5:i64 = Load v4, 0x8
-	v6:i64 = Load module_ctx, 0x18
-	v7:f32 = Load v6, 0x8
-	v8:i64 = Load module_ctx, 0x20
-	v9:f64 = Load v8, 0x8
+	v2:i32 = Load module_ctx, 0x8
+	v3:i64 = Load module_ctx, 0x18
+	v4:f32 = Load module_ctx, 0x28
+	v5:f64 = Load module_ctx, 0x38
 	Store module_ctx, exec_ctx, 0x8
 	Call f1:sig1, exec_ctx, module_ctx
-	v10:i64 = Load module_ctx, 0x8
-	v11:i32 = Load v10, 0x8
-	v12:i64 = Load module_ctx, 0x10
-	v13:i64 = Load v12, 0x8
-	v14:i64 = Load module_ctx, 0x18
-	v15:f32 = Load v14, 0x8
-	v16:i64 = Load module_ctx, 0x20
-	v17:f64 = Load v16, 0x8
-	Jump blk_ret, v3, v5, v7, v9, v11, v13, v15, v17
+	v6:i32 = Load module_ctx, 0x8
+	v7:i64 = Load module_ctx, 0x18
+	v8:f32 = Load module_ctx, 0x28
+	v9:f64 = Load module_ctx, 0x38
+	Jump blk_ret, v2, v3, v4, v5, v6, v7, v8, v9
 `,
 			expAfterOpt: `
 signatures:
 	sig1: i64i64_v
 
 blk0: (exec_ctx:i64, module_ctx:i64)
-	v2:i64 = Load module_ctx, 0x8
-	v3:i32 = Load v2, 0x8
-	v4:i64 = Load module_ctx, 0x10
-	v5:i64 = Load v4, 0x8
-	v6:i64 = Load module_ctx, 0x18
-	v7:f32 = Load v6, 0x8
-	v8:i64 = Load module_ctx, 0x20
-	v9:f64 = Load v8, 0x8
+	v2:i32 = Load module_ctx, 0x8
+	v3:i64 = Load module_ctx, 0x18
+	v4:f32 = Load module_ctx, 0x28
+	v5:f64 = Load module_ctx, 0x38
 	Store module_ctx, exec_ctx, 0x8
 	Call f1:sig1, exec_ctx, module_ctx
-	v10:i64 = Load module_ctx, 0x8
-	v11:i32 = Load v10, 0x8
-	v12:i64 = Load module_ctx, 0x10
-	v13:i64 = Load v12, 0x8
-	v14:i64 = Load module_ctx, 0x18
-	v15:f32 = Load v14, 0x8
-	v16:i64 = Load module_ctx, 0x20
-	v17:f64 = Load v16, 0x8
-	Jump blk_ret, v3, v5, v7, v9, v11, v13, v15, v17
+	v6:i32 = Load module_ctx, 0x8
+	v7:i64 = Load module_ctx, 0x18
+	v8:f32 = Load module_ctx, 0x28
+	v9:f64 = Load module_ctx, 0x38
+	Jump blk_ret, v2, v3, v4, v5, v6, v7, v8, v9
 `,
 		},
 		{