--- a/internal/wasm/global.go
+++ b/internal/wasm/global.go
@@ -18,7 +18,8 @@ func (g constantGlobal) Type() api.ValueType {
 
 // Get implements api.Global.
 func (g constantGlobal) Get() uint64 {
-	return g.g.Val
+	ret, _ := g.g.Value()
+	return ret
 }
 
 // String implements api.Global.
@@ -39,7 +40,8 @@ func (g mutableGlobal) Type() api.ValueType {
 
 // Get implements api.Global.
 func (g mutableGlobal) Get() uint64 {
-	return g.g.Val
+	ret, _ := g.g.Value()
+	return ret
 }
 
 // String implements api.Global.
@@ -47,10 +49,7 @@ func (g mutableGlobal) String() string {
 	return g.g.String()
 }
 
-// Set implements the same method as documented on api.MutableGlobal.
+// Set implements the same method as documented on api.mutableGlobal.
 func (g mutableGlobal) Set(v uint64) {
 	g.g.Val = v
 }
-
-// compile-time check to ensure mutableGlobal is a api.Global.
-var _ api.Global = mutableGlobal{}