… internal/engine/wazevo/module_engine_test.go
… internal/engine/wazevo/wazevoapi/offsetdata_test.go
… internal/integration_test/engine/adhoc_test.go
… internal/integration_test/fuzzcases/fuzzcases_test.go
… internal/wasm/store_test.go