@mathetake 2023-12-22
found a bug not covered by the existing ad_hoc test so I need to add one, and after that, will run fuzzing